Move evaluation script from load_agent.py into separate script
