+++ README.md
@@ -74,12 +74,20 @@ obviously don't need to create a new one.
 
 * To evaluate a previously trained agent on a dataset, you may use the `evaluate` method available as a click CLI when executing:
     ```bash
-    $ python load_agent.py
+    $ python evaluate_agent.py
     ```
-    (Run `python load_agent.py --help` to see the required parameters for the CLI)
+    (Run `python evaluate_agent.py --help` to see the required parameters for the CLI)
 * If you provide the `--save` flag in the CLI above, it creates `.npy` files which can be read by the `evaluate_from_files` CLI afterwards:
     ```bash
     $ python evaluate_from_files.py
     ```
     (Run `python evaluate_from_files.py --help` to see the required parameters for the CLI)
 * The `evaluate_from_files` CLI allows defining an IoU threshold used for the calculation of the evaluation metrics. Furthermore, it does not only output the mean average precision (mAP) but also the precision and recall values.
+
+## Creating image sequences/animations for visualization purposes
+
+* To create an image sequence of a an already trained agent acting on a specific image, use:
+    ```bash
+    $ python generate_image_sequence.py
+    ```
+    (Run `python generate_image_sequence.py --help` to see the required parameters for the CLI and have a look into the `generate_image_sequence.py` file for instructions on creating a video out of the generated single frames using ffmpeg)
